Move General to workspaces folder
