+++ packages/frontend-2/components/settings/workspaces/General/DeleteDialog.vue
